+++ packages/tideline/js/tidelinedata.js
@@ -455,12 +455,12 @@ function TidelineData(data, opts) {
     this.setDeviceParameters(this.data);
 
     this.zenEvents = this.setEvents(
-      this.data, 
+      this.data,
       {type: 'deviceEvent', subType: 'zen'},
       ['inputTime']
     );
     this.confidentialEvents = this.setEvents(
-      this.data, 
+      this.data,
       {type: 'deviceEvent', subType: 'confidential'},
       ['inputTime']
     );
@@ -659,8 +659,13 @@ function TidelineData(data, opts) {
           d.localDayOfWeek = dt.getLocalDayOfWeek(date, opts.timePrefs.timezoneName);
           d.localDate = dt.getLocalDate(date, opts.timePrefs.timezoneName);
           d.msPer24 = dt.getMsPer24(d.normalTime, opts.timePrefs.timezoneName);
+          }
+        if (d.type === 'deviceEvent' && d.subType === 'confidential') {
+          const m = moment.utc(d.time);
+          m.add(_.get(d, 'duration.value', 0), _.get(d, 'duration.units', 'hours'));
+          d.normalEnd = m.toISOString();
         }
-      };
+    };
     }
     else {
       watson = function(d) {
@@ -788,12 +793,12 @@ function TidelineData(data, opts) {
   endTimer('deduplicatePhysicalActivities');
 
   this.zenEvents = this.setEvents(
-    data, 
+    data,
     {type: 'deviceEvent', subType: 'zen'},
     ['inputTime']
   );
   this.confidentialEvents = this.setEvents(
-    data, 
+    data,
     {type: 'deviceEvent', subType: 'confidential'},
     ['inputTime']
   );